rename: 프로필관련 컴포넌트 디렉토리 재분류류
