+++ app/profile/(sniped)/SnipedPostList.tsx
