+++ app/profile/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 
 import PlatformProfiles from "#app/profile/(platform)/PlatformProfiles.jsx";
-import SnipedPostListSection from "#app/profile/(platform)/SnipedPostListSection.jsx";
+import SnipedPostListSection from "#app/profile/(sniped)/SnipedPostListSection.jsx";
 import { auth } from "#auth";
 import { getProfile } from "#lib/database/users.js";
